Fix gcc.dg/warn-abs-1.c for arm and aarch64-none-elf
This new test has some difficulties on the fabsl function.
On arm this is because we don't support the _Float128 type which the test uses.
This is handled in the patch by requiring a float128 target selector.

On aarch64-none-elf, a Newlib target, it fails because fabsl is not available.
long double support is known to be incomplete in newlib, and the fabsl function is not available
for targets where long double is larger than a double.
Therefore this patch skips the test on such targets.


	* gcc.dg/warn-abs-1.c: Require float128 target.
	Skip if large_long_double newlib target.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@264392 138bc75d-0d04-0410-961f-82ee72b054a4