Remove superfluous assignment in add_params.
This fixes the superfluous assignment that Coverity reported in add_params,
and changes the starting index from 0 to num_params - n in order for it to
work properly if add_params is called multiple times.

validate_params calls error so it doesn't matter that we don't check the
results here.  The results is checked in individual parameter updates after
front-end initialization.

2018-10-04  Tamar Christina  <tamar.christina@arm.com>

	* params.c (add_params): Fix initialization.



git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@264841 138bc75d-0d04-0410-961f-82ee72b054a4